Move generateFeedbackUrl test from footer_tests to utils_tests
